chore(refactor): file file and folder structure
