+++ client/package.json
@@ -20,7 +20,6 @@
     "@types/react-dom": "^18.0.6",
     "@types/styled-components": "^5.1.26",
     "@vitejs/plugin-react": "^2.1.0",
-    "@vitejs/plugin-react-refresh": "^1.3.6",
     "jest": "^29.1.2",
     "jest-environment-jsdom": "^29.1.2",
     "lerna": "^5.6.2",
