+++ client/packages/portal-client/src/components/phase-icons/Early.tsx
