+++ client/packages/portal-client/src/components/phase-icons/Execution.tsx
