+++ client/packages/portal-client/src/components/phase-icons/Renewable.tsx
