+++ client/packages/portal-pages/src/pages/home-page/HomePageIcon.tsx
