+++ client/packages/portal-ui/src/index.ts
@@ -1,6 +1,6 @@
-export * from './lib/portalHeader/Icon';
-export * from './lib/portalHeader/PortalHeader';
-export * from './lib/portalMenu/MenuButton';
-export * from './lib/portalMenu/MenuContext';
-export * from './lib/portalMenu/PortalMenu';
-export * from './lib/styleProvider/styleProvider';
+export * from './lib/portal-header/Icon';
+export * from './lib/portal-header/PortalHeader';
+export * from './lib/portal-menu/MenuButton';
+export * from './lib/portal-menu/MenuContext';
+export * from './lib/portal-menu/PortalMenu';
+export * from './lib/style-provider/styleProvider';
