+++ client/packages/portal-ui/src/lib/portal-header/Logo.tsx
