+++ client/packages/portal-ui/src/lib/style-provider/globalColors.ts
